Changes node count to type int.

+++ JAICore/jaicore-search/src/jaicore/search/algorithms/standard/uncertainty/paretosearch/ParetoSelection.java
@@ -19,7 +19,7 @@ public class ParetoSelection <T, V extends Comparable<V>> implements OpenCollect
 	private final Queue<ParetoNode<T,V>> pareto;
 
 	/* Node counter. */
-	private long n = 0;
+	private int n = 0;
 
 	/**
 	 * Constructor.
